# Improve performance of allocation profiling