# What does this PR do?

- Adjust the pseudo random generator to be thread local
- Change the generator to be smaller
- Remove the lock from the allocation profiling path

Credit for this idea goes to @nsavoire 

# Motivation

Improve performance of allocation profiling. This really brings down the allocation profiling to numbers that are comparable. 
```
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_ThreadedAllocations_NoTracking     343085 ns       172913 ns         4525
BM_ThreadedAllocations_Tracking       483263 ns       244706 ns         2886
```

The numbers were 4x worse prior to this change.

After this change I will need to focus on the deallocation code path.

# Additional Notes

This needs extensive testing.
Risks are:
- Accuracy 
- Thread safety

# How to test the change?

The benchmark only tests parts of this. We will need to be careful when deploying this change.